Lesson_31_Jenkins/qauto/requirements.txt
pytest==8.3.4
playwright==1.49.1
allure-pytest==2.13.5
requests==2.32.3